+++ packages/orm/CHANGELOG.md
@@ -1,5 +1,12 @@
 # orchid-orm
 
+## 1.3.12
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.7.9
+
 ## 1.3.11
 
 ### Patch Changes