+++ packages/test-factory/CHANGELOG.md
@@ -1,5 +1,14 @@
 # orchid-orm-test-factory
 
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.7.9
+  - orchid-orm@1.3.12
+  - orchid-orm-schema-to-zod@0.1.9
+
 ## 0.1.11
 
 ### Patch Changes